+++ tests-integration/tests/validator_staking.rs
@@ -3,7 +3,6 @@
 use andromeda_app::app::AppComponent;
 use andromeda_app_contract::mock::{mock_andromeda_app, MockAppContract};
 
-use andromeda_std::amp::AndrAddr;
 use andromeda_testing::mock::mock_app;
 use andromeda_testing::mock_builder::MockAndromedaBuilder;
 use andromeda_validator_staking::mock::{
@@ -11,7 +10,7 @@ use andromeda_validator_staking::mock::{
 };
 
 // use andromeda_std::error::ContractError;
-use andromeda_std::error::ContractError::Std;
+use andromeda_std::error::ContractError::{self, Std};
 use andromeda_testing::MockContract;
 use cosmwasm_std::StdError::GenericErr;
 use cosmwasm_std::{coin, to_json_binary, Addr, BlockInfo, Delegation, Uint128};
@@ -28,7 +27,6 @@ fn test_validator_stake() {
         ])
         .build(&mut router);
     let owner = andr.get_wallet("owner");
-    let recipient = AndrAddr::from_string(owner.to_string());
     let validator_1 = router.api().addr_make("validator1");
 
     let validator_staking_init_msg = mock_validator_staking_instantiate_msg(
@@ -71,12 +69,7 @@ fn test_validator_stake() {
     // Testing when there is no reward to claim
     // TODO: These errors cant be downcast anymore?
     let _err = validator_staking
-        .execute_claim_reward(
-            &mut router,
-            owner.clone(),
-            Some(validator_1.clone()),
-            Some(recipient.clone()),
-        )
+        .execute_claim_reward(&mut router, owner.clone(), Some(validator_1.clone()))
         .unwrap_err();
     // assert_eq!(may_err.unwrap(), &expected_err);
 
@@ -90,32 +83,17 @@ fn test_validator_stake() {
         chain_id: router.block_info().chain_id,
     });
 
-    // only owner can become a recipient
-    let _err = validator_staking
-        .execute_claim_reward(
-            &mut router,
-            owner.clone(),
-            Some(validator_1.clone()),
-            Some(AndrAddr::from_string("some_address")),
-        )
-        .unwrap_err();
-    // let _err = err.root_cause().downcast_ref::<ContractError>().unwrap();
-    // let expected_err = ContractError::Unauthorized {};
-    // assert_eq!(err, &expected_err);
-
     validator_staking
-        .execute_claim_reward(
-            &mut router,
-            owner.clone(),
-            Some(validator_1),
-            Some(recipient),
-        )
+        .execute_claim_reward(&mut router, owner.clone(), Some(validator_1))
         .unwrap();
 
     // Default APR 10% by cw-multi-test -> StakingInfo
     // should now have 1000 * 10% / 2 - 0% commission = 50 tokens reward
-    let owner_balance = router.wrap().query_balance(owner.clone(), "TOKEN").unwrap();
-    assert_eq!(owner_balance, coin(50, "TOKEN"));
+    let contract_balance = router
+        .wrap()
+        .query_balance(validator_staking.addr(), "TOKEN")
+        .unwrap();
+    assert_eq!(contract_balance, coin(50, "TOKEN"));
 
     // Test unstake with invalid validator
     let _err = validator_staking
@@ -160,16 +138,6 @@ fn test_validator_stake() {
         })
     );
 
-    // Test withdraw before payout period
-    let _err = validator_staking
-        .execute_withdraw_fund(&mut router, owner.clone())
-        .unwrap_err();
-    // let _err = err.root_cause().downcast_ref::<ContractError>().unwrap();
-    // let expected_err = ContractError::InvalidWithdrawal {
-    //     msg: Some("No unstaked funds to withdraw".to_string()),
-    // };
-    // assert_eq!(err, &expected_err);
-
     let unstaked_tokens = validator_staking.query_unstaked_tokens(&router).unwrap();
     let unbonding_period =
         unstaked_tokens[0].payout_at.seconds() - router.block_info().time.seconds();
@@ -206,7 +174,6 @@ fn test_validator_stake_and_unstake_specific_amount() {
         ])
         .build(&mut router);
     let owner = andr.get_wallet("owner");
-    let recipient = AndrAddr::from_string(owner.to_string());
     let validator_1 = router.api().addr_make("validator1");
 
     let validator_staking_init_msg = mock_validator_staking_instantiate_msg(
@@ -248,12 +215,7 @@ fn test_validator_stake_and_unstake_specific_amount() {
 
     // Testing when there is no reward to claim
     let _err = validator_staking
-        .execute_claim_reward(
-            &mut router,
-            owner.clone(),
-            Some(validator_1.clone()),
-            Some(recipient.clone()),
-        )
+        .execute_claim_reward(&mut router, owner.clone(), Some(validator_1.clone()))
         .unwrap_err();
     // assert_eq!(may_err.unwrap(), &expected_err);
 
@@ -267,29 +229,17 @@ fn test_validator_stake_and_unstake_specific_amount() {
         chain_id: router.block_info().chain_id,
     });
 
-    // only owner can become a recipient
-    let _err = validator_staking
-        .execute_claim_reward(
-            &mut router,
-            owner.clone(),
-            Some(validator_1.clone()),
-            Some(AndrAddr::from_string("some_address")),
-        )
-        .unwrap_err();
-
     validator_staking
-        .execute_claim_reward(
-            &mut router,
-            owner.clone(),
-            Some(validator_1),
-            Some(recipient),
-        )
+        .execute_claim_reward(&mut router, owner.clone(), Some(validator_1))
         .unwrap();
 
     // Default APR 10% by cw-multi-test -> StakingInfo
     // should now have 1000 * 10% / 2 - 0% commission = 50 tokens reward
-    let owner_balance = router.wrap().query_balance(owner.clone(), "TOKEN").unwrap();
-    assert_eq!(owner_balance, coin(50, "TOKEN"));
+    let contract_balance = router
+        .wrap()
+        .query_balance(validator_staking.addr(), "TOKEN")
+        .unwrap();
+    assert_eq!(contract_balance, coin(50, "TOKEN"));
 
     // Test unstake with invalid validator
     let _err = validator_staking
@@ -341,14 +291,10 @@ fn test_validator_stake_and_unstake_specific_amount() {
         }
     );
 
-    // Test withdraw before payout period
-    let _err = validator_staking
-        .execute_withdraw_fund(&mut router, owner.clone())
-        .unwrap_err();
-
     let unstaked_tokens = validator_staking.query_unstaked_tokens(&router).unwrap();
     let unbonding_period =
         unstaked_tokens[0].payout_at.seconds() - router.block_info().time.seconds();
+
     // Update block to payout period
     router.set_block(BlockInfo {
         height: router.block_info().height,
@@ -368,4 +314,72 @@ fn test_validator_stake_and_unstake_specific_amount() {
 
     let owner_balance = router.wrap().query_balance(owner, "TOKEN").unwrap();
     assert_eq!(owner_balance, coin(250, "TOKEN"));
+}
+
+#[test]
+fn test_update_default_validator() {
+    let mut router = mock_app(Some(vec!["TOKEN"]));
+
+    let andr = MockAndromedaBuilder::new(&mut router, "admin")
+        .with_wallets(vec![("owner", vec![coin(1000, "TOKEN")])])
+        .with_contracts(vec![
+            ("app-contract", mock_andromeda_app()),
+            ("validator-staking", mock_andromeda_validator_staking()),
+        ])
+        .build(&mut router);
+    let owner = andr.get_wallet("owner");
+    let validator_1 = router.api().addr_make("validator1");
+    let validator_2 = router.api().addr_make("validator2");
+    let validator_3 = router.api().addr_make("validator3");
+
+    let validator_staking_init_msg = mock_validator_staking_instantiate_msg(
+        validator_1.clone(),
+        None,
+        andr.kernel.addr().to_string(),
+    );
+
+    let validator_staking_component = AppComponent::new(
+        "staking".to_string(),
+        "validator-staking".to_string(),
+        to_json_binary(&validator_staking_init_msg).unwrap(),
+    );
+
+    let app_components = vec![validator_staking_component.clone()];
+    let app = MockAppContract::instantiate(
+        andr.get_code_id(&mut router, "app-contract"),
+        owner,
+        &mut router,
+        "Validator Staking App",
+        app_components,
+        andr.kernel.addr(),
+        Some(owner.to_string()),
+    );
+
+    let validator_staking: MockValidatorStaking =
+        app.query_ado_by_component_name(&router, validator_staking_component.name);
+
+    // Update default validator with invalid validator
+    let err: ContractError = validator_staking
+        .execute_update_default_validator(&mut router, owner.clone(), validator_3.clone())
+        .unwrap_err()
+        .downcast()
+        .unwrap();
+    assert_eq!(err, ContractError::InvalidValidator {});
+
+    // Update default validator
+    validator_staking
+        .execute_update_default_validator(&mut router, owner.clone(), validator_2.clone())
+        .unwrap();
+
+    let funds = vec![coin(1000, "TOKEN")];
+
+    // Stake with default validator
+    validator_staking
+        .execute_stake(&mut router, owner.clone(), None, funds)
+        .unwrap();
+
+    let stake_info = validator_staking
+        .query_staked_tokens(&router, None)
+        .unwrap();
+    assert_eq!(stake_info.validator, validator_2.to_string());
 }